Fix synchronization issue on join
Here the client did not wait long enough. This meant that the server was
still resetting the queues which could drop requests by a client by
storing 0 in the write_head, which was assumed to be controlled by the
newly joining client.

The issue was that the client accepted _any_ waking from the futex as
sufficient signal that the server had looked at its request. But indeed,
a second client that joins and blocks might also wake it instead. We
now utilize the bitset feature which requires us to only support
`i32::MAX` clients at most. The highest bit is used to signal the empty
slot state. This allows us to use wait_bitset with which we
differentiate between bad wakes and successful wakes.