Configure more than one page per queue
The PAGE_SIZE is given by hardware, we have little to no influence on it
and can't directly increase it. But we want our queues to have room for
a good lot of data so we will allocate more than a single for each
direction and each scratch space, ensuring that 1<<16 bytes are
available in either.