Fix waiting being switched
It's so unclear how to use timeouts. Sure, we need OP_LINK_TIMEOUT but
the documentation is not discoverable. At all. An older thread [1]
spells it out but without indicating any responsible function so I
fucking used OP_TIMEOUT for it. The documentation has nothing to say.
[2] Not even that we can cancel other request with this. Fuck it. The
travesty is that the liburing documentation is the primary one anyways,
like are we treating mailing archives as that now? Need more kernel devs
working on Rust to realize just *how* bad the situation and compare them
to good API explanations.

Anyways, the timeout now effectively works. Then also we make sure the
poll's read readiness is cleared so we don't go into a constant loop.

[1]: https://lwn.net/Articles/803932/
[2]: https://web.archive.org/web/20240515200724/https://man.archlinux.org/man/io_uring_prep_timeout.3.en