Ensure re-initialization of rings with one side
If a client takes one side of a ring to offer a service and then dies,
the implementation never recognized this as a ring to bring up. It would
go into a state where the other side was still shared but would only
re-initialize rings where both are down.

The new code will take over rings by the server where one side is dead,
if the other is still being shared. It can then re-initialize the queues
and bring them up again.